python/solution_tasks/sdr_projects_2025/requirements.txt
numpy
matplotlib
pyrtlsdr
jax
requests
hackrf
scipy
